Updated Objectives
I updated the profiles for objectives based on the "End of Round" and
"End of Game" categories no longer being limited to certain types of
objectives.  I also centralized the objectives in the gst, set defaults
(first alphabetically) to avoid selecting the category but never picking
a specific objective, and added the missing CC content.